changed public API LJUSB_GetDevCount to return unsigned int instead of unsigned long. It was in fact returning unsigned int really anyway. This fixes the need to cast in client code.

+++ liblabjackusb/labjackusb.c
@@ -899,7 +899,7 @@ void LJUSB_CloseDevice(HANDLE hDevice)
 }
 
 
-unsigned long LJUSB_GetDevCount(unsigned long ProductID)
+unsigned int LJUSB_GetDevCount(unsigned long ProductID)
 {
     libusb_device **devs;
     ssize_t cnt;
